+++ src/trx/game/sparks/spawners.h
@@ -25,6 +25,8 @@ void Sparks_TriggerFireSmoke(XYZ_32 pos, int32_t body_part, int32_t type);
 void Sparks_TriggerSideFlame(
     XYZ_32 pos, int32_t angle, int32_t speed, bool pilot);
 
+void Sparks_TriggerDartSmoke(XYZ_32 pos, XZ_32 vel, bool hit);
+
 void Sparks_TriggerFlareSparks(XYZ_32 pos, XYZ_32 vel, bool smoke);
 
 void Sparks_TriggerRicochet(GAME_VECTOR pos, int32_t angle, int32_t size);